workflow/requirements.txt
google-cloud-workflows>=0.1.0
anyio==3.6.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
docopt==0.6.2
fastapi==0.82.0
flask>=1.0.2
google-api-core==2.10.0
google-auth==2.11.0
google-cloud==0.34.0
google-cloud-bigquery==3.3.2
google-cloud-bigquery-storage==2.14.2
google-cloud-bigtable==2.11.3
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.48.1
grpcio-status==1.48.1
h11==0.13.0
idna==3.3
numpy>=1.7.2
packaging==21.3
pipreqs==0.4.11
proto-plus==1.22.1
protobuf==4.21.5
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
yarg==0.1.9